Added date sorting to save screen.
[Start SVN Log Rant]
Unfortunately there's some annoying autosave name bug.  Right now there is an autosave set to the cockpit's unitmodname member which is the config var "player","callsign".
However in other places it is set to outputsavename variable callsign+"_"+mission's "savegame" var.
This file generation name is pretty arbitrary and will change if you play multiplayer since it uses the callsign, and it is copied to another name anyway.