Implement degraded governor
This seems to be the most elegant way of re-creating the degraded Llama that the game is started with. 

Removed the Overdrive governor so that the ship gets the default speed for overdrive. It is highly unlikely that you would purchase an overdrive before doing at least the first repair and refuel. Even if you do, the effect on game play is less than forever being stuck with an overdrive that is less performant than the default Llama. 

Tested with a new campaign, and before purchasing a repair and refuel the governor speed is 120. After the repair and refuel the governor speed has changed to 125.