Update config.json
This variable has been calibrated to make the Llama match its 6.9 gravities forward acceleration.

Please test with your own ships. 
Vessel speed is measured in knots, and check the acceleration listed in the ship info in the base computer. 

For the Llama this is 6.9g forward acelleration. 
With this acceleration, at 100% weight, it takes 30 seconds to reach 4000kps.